requirements.txt
# Core Dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

# Google Gemini
google-generativeai==0.3.2

# LangChain
langchain==0.1.7
langchain-community==0.0.20

# Vector Store
faiss-cpu==1.13.2
numpy==1.26.3

# PDF Processing
pymupdf==1.23.21
PyPDF2==3.0.1

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
aiofiles==23.2.1

# CORS
python-jose[cryptography]==3.3.0
